+++ android/src/main/java/com/easemob/im_flutter_sdk/EMChatManagerWrapper.java
@@ -46,8 +46,6 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
     }
 
 
-
-
     @Override
     public void onMethodCall(MethodCall call, Result result) {
         JSONObject param = (JSONObject) call.arguments;
@@ -239,6 +237,7 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
         asyncRunnable(()->{
             try {
                 EMClient.getInstance().chatManager().ackGroupMessageRead(to, msgId, finalContent);
+                onSuccess(result, channelName, true);
             } catch (HyphenateException e) {
                 onError(result, e);
             }
@@ -280,7 +279,11 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
 
         asyncRunnable(() -> {
             EMMessage msg = EMClient.getInstance().chatManager().getMessage(msgId);
-            onSuccess(result, channelName, EMMessageHelper.toJson(msg));
+            if(msg == null) {
+                onSuccess(result, channelName, null);
+            }else {
+                onSuccess(result, channelName, EMMessageHelper.toJson(msg));
+            }
         });
     }
 
@@ -436,9 +439,13 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
     }
 
     private void loadAllConversations(JSONObject param, String channelName, Result result) throws JSONException {
+        if (EMClient.getInstance().getCurrentUser() == null || EMClient.getInstance().getCurrentUser().length() == 0) {
+            onSuccess(result, channelName, new ArrayList<>());
+            return;
+        }
+        List<EMConversation> list = new ArrayList<>(EMClient.getInstance().chatManager().getAllConversations().values());
         asyncRunnable(() -> {
             boolean retry = false;
-            List<EMConversation> list = new ArrayList<>(EMClient.getInstance().chatManager().getAllConversations().values());
             List<Map> conversations = new ArrayList<>();
             do{
                 try{
@@ -536,12 +543,12 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
 
     private void searchChatMsgFromDB(JSONObject param, String channelName, Result result) throws JSONException {
         String keywords = param.getString("keywords");
-        long timeStamp = param.getLong("timeStamp");
+        long timestamp = param.getLong("timestamp");
         int count = param.getInt("maxCount");
         String from = param.getString("from");
         EMSearchDirection direction = searchDirectionFromString(param.getString("direction"));
         asyncRunnable(() -> {
-            List<EMMessage> msgList = EMClient.getInstance().chatManager().searchMsgFromDB(keywords, timeStamp, count,
+            List<EMMessage> msgList = EMClient.getInstance().chatManager().searchMsgFromDB(keywords, timestamp, count,
                     from, direction);
             List<Map> messages = new ArrayList<>();
             for (EMMessage msg : msgList) {
@@ -772,8 +779,9 @@ public class EMChatManagerWrapper extends EMWrapper implements MethodCallHandler
         };
 
         conversationListener = new EMConversationListener() {
+
             @Override
-            public void onCoversationUpdate() {
+            public void onConversationUpdate() {
                 Map<String, Object> data = new HashMap<>();
                 post(() -> channel.invokeMethod(EMSDKMethod.onConversationUpdate, data));
             }