+++ lib/src/internal/chat_method_keys.dart
@@ -1,3 +1,4 @@
+/// @nodoc
 class ChatMethodKeys {
   /// EMClient methods
   static const String init = "init";
@@ -179,6 +180,10 @@ class ChatMethodKeys {
 
   static const String muteAllChatRoomMembers = "muteAllChatRoomMembers";
   static const String unMuteAllChatRoomMembers = "unMuteAllChatRoomMembers";
+
+  static const String fetchChatRoomAttributes = "fetchChatRoomAttributes";
+  static const String setChatRoomAttributes = "setChatRoomAttributes";
+  static const String removeChatRoomAttributes = "removeChatRoomAttributes";
 
   /// EMChatRoomManagerListener
   static const String chatRoomChange = "onChatRoomChanged";