+++ release.sh
@@ -1,0 +1,2 @@
+flutter format .
+flutter packages pub publish --server=https://pub.dartlang.org